backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==2.6.4
google-api-python-client==2.118.0
python-dotenv
libsql-client
mangum
opensearch-py==3.1.0
